+++ PhpSerializerNET/Deserializers/Array/UntypedArrayDeserializer.cs
@@ -14,23 +14,20 @@ internal class UntypedArrayDeserializer : ArrayDeserializer {
 	}
 
 	internal override object Deserialize(PhpSerializeToken token) {
-		switch (token.Type) {
-			case PhpSerializerType.Array:
-				return this.DeserializeArray(token);
-			case PhpSerializerType.Object:
-				return this.ObjectDeserializer.Deserialize(token);
-			default:
-				return this.PrimitiveDeserializer.Deserialize(token);
-		}
+		return token.Type switch {
+			PhpSerializerType.Array => this.DeserializeArray(token),
+			PhpSerializerType.Object => this.ObjectDeserializer.Deserialize(token),
+			_ => this.PrimitiveDeserializer.Deserialize(token),
+		};
 	}
 
 	public object DeserializeArray(PhpSerializeToken token) {
 		if (this._options.UseLists == ListOptions.Never) {
 			var result = new Dictionary<object, object>();
-			for (int i = 0; i < token.Children.Length; i += 2) {
+			foreach(var item in token.Children) {
 				result.Add(
-					this.Deserialize(token.Children[i]),
-					this.Deserialize(token.Children[i + 1])
+					this.Deserialize(item.Key),
+					this.Deserialize(item.Value)
 				);
 			}
 			return result;
@@ -38,12 +35,13 @@ internal class UntypedArrayDeserializer : ArrayDeserializer {
 		long previousKey = -1;
 		bool isList = true;
 		bool consecutive = true;
-		for (int i = 0; i < token.Children.Length; i += 2) {
-			if (token.Children[i].Type != PhpSerializerType.Integer) {
+		for (int i = 0; i < token.Children.Length; i++) {
+			var item = token.Children[i];
+			if (item.Key.Type != PhpSerializerType.Integer) {
 				isList = false;
 				break;
 			} else {
-				var key = token.Children[i].Value.PhpToLong();
+				var key = item.Key.Value.PhpToLong();
 				if (i == 0 || key == previousKey + 1) {
 					previousKey = key;
 				} else {
@@ -53,17 +51,17 @@ internal class UntypedArrayDeserializer : ArrayDeserializer {
 		}
 		if (!isList || (this._options.UseLists == ListOptions.Default && consecutive == false)) {
 			var result = new Dictionary<object, object>();
-			for (int i = 0; i < token.Children.Length; i += 2) {
+			foreach(var item in token.Children) {
 				result.Add(
-					this.Deserialize(token.Children[i]),
-					this.Deserialize(token.Children[i + 1])
+					this.Deserialize(item.Key),
+					this.Deserialize(item.Value)
 				);
 			}
 			return result;
 		} else {
 			var result = new List<object>();
-			for (int i = 1; i < token.Children.Length; i += 2) {
-				result.Add(this.Deserialize(token.Children[i]));
+			foreach(var item in token.Children) {
+				result.Add(this.Deserialize(item.Value));
 			}
 			return result;
 		}