+++ PhpSerializerNET/Deserializers/Object/TypedObjectDeserializer.cs
@@ -66,10 +66,10 @@ internal class TypedObjectDeserializer : ObjectDeserializer {
 			} else {
 				throw new DeserializationException("Encountered 'stdClass' and the behavior 'Throw' was specified in deserialization options.");
 			}
-			for (int i = 0; i < token.Children.Length; i += 2) {
+			foreach(var item in token.Children) {
 				result.TryAdd(
-					token.Children[i].Value,
-					this.Deserialize(token.Children[i + 1])
+					item.Key.Value,
+					this.Deserialize(item.Value)
 				);
 			}
 			constructedObject = result;
@@ -84,25 +84,22 @@ internal class TypedObjectDeserializer : ObjectDeserializer {
 		var result = Activator.CreateInstance(targetType);
 		Dictionary<object, PropertyInfo> properties = TypeLookup.GetPropertyInfos(targetType, this._options);
 
-		for (int i = 0; i < token.Children.Length; i += 2) {
+		foreach(var item in token.Children) {
 			object propertyName;
-			if (token.Children[i].Type == PhpSerializerType.String) {
-				propertyName = this._options.CaseSensitiveProperties ? token.Children[i].Value : token.Children[i].Value.ToLower();
-			} else if (token.Children[i].Type == PhpSerializerType.Integer) {
-				propertyName = token.Children[i].Value.PhpToLong();
+			if (item.Key.Type == PhpSerializerType.String) {
+				propertyName = this._options.CaseSensitiveProperties ? item.Key.Value : item.Key.Value.ToLower();
+			} else if (item.Key.Type == PhpSerializerType.Integer) {
+				propertyName = item.Key.Value.PhpToLong();
 			} else {
 				throw new DeserializationException(
 					$"Error encountered deserizalizing an object of type '{targetType.FullName}': " +
-					$"The key '{token.Children[i].Value}' (from the token at position {token.Children[i].Position}) has an unsupported type of '{token.Children[i].Type}'."
+					$"The key '{item.Key.Value}' (from the token at position {item.Key.Position}) has an unsupported type of '{item.Key.Type}'."
 				);
 			}
-
-			var valueToken = token.Children[i + 1];
-
 			if (!properties.ContainsKey(propertyName)) {
 				if (!this._options.AllowExcessKeys) {
 					throw new DeserializationException(
-						$"Could not bind the key \"{token.Children[i].Value}\" to object of type {targetType.Name}: No such property."
+						$"Could not bind the key \"{item.Key.Value}\" to object of type {targetType.Name}: No such property."
 					);
 				}
 				continue;
@@ -112,11 +109,11 @@ internal class TypedObjectDeserializer : ObjectDeserializer {
 				try {
 					property.SetValue(
 						result,
-						this.Deserialize(valueToken, property.PropertyType)
+						this.Deserialize(item.Value, property.PropertyType)
 					);
 				} catch (Exception exception) {
 					throw new DeserializationException(
-						$"Exception encountered while trying to assign '{valueToken.Value}' to {targetType.Name}.{property.Name}. See inner exception for details.",
+						$"Exception encountered while trying to assign '{item.Value.Value}' to {targetType.Name}.{property.Name}. See inner exception for details.",
 						exception
 					);
 				}
@@ -129,13 +126,12 @@ internal class TypedObjectDeserializer : ObjectDeserializer {
 		var result = Activator.CreateInstance(targetType);
 		Dictionary<string, FieldInfo> fields = TypeLookup.GetFieldInfos(targetType, this._options);
 
-		for (int i = 0; i < token.Children.Length; i += 2) {
-			var fieldName = this._options.CaseSensitiveProperties ? token.Children[i].Value : token.Children[i].Value.ToLower();
-			var valueToken = token.Children[i + 1];
+		foreach(var item in token.Children) {
+			var fieldName = this._options.CaseSensitiveProperties ? item.Key.Value : item.Key.Value.ToLower();
 			if (!fields.ContainsKey(fieldName)) {
 				if (!this._options.AllowExcessKeys) {
 					throw new DeserializationException(
-						$"Could not bind the key \"{token.Children[i].Value}\" to struct of type {targetType.Name}: No such field."
+						$"Could not bind the key \"{item.Key.Value}\" to struct of type {targetType.Name}: No such field."
 					);
 				}
 				continue;
@@ -143,10 +139,10 @@ internal class TypedObjectDeserializer : ObjectDeserializer {
 			if (fields[fieldName] != null) {
 				var field = fields[fieldName];
 				try {
-					field.SetValue(result, this.Deserialize(valueToken, field.FieldType));
+					field.SetValue(result, this.Deserialize(item.Value, field.FieldType));
 				} catch (Exception exception) {
 					throw new DeserializationException(
-						$"Exception encountered while trying to assign '{valueToken.Value}' to {targetType.Name}.{field.Name}. " +
+						$"Exception encountered while trying to assign '{item.Value}' to {targetType.Name}.{field.Name}. " +
 						"See inner exception for details.",
 						exception
 					);
@@ -162,12 +158,10 @@ internal class TypedObjectDeserializer : ObjectDeserializer {
 			throw new NullReferenceException($"Activator.CreateInstance({targetType.FullName}) returned null");
 		}
 		if (!targetType.GenericTypeArguments.Any()) {
-			for (int i = 0; i < token.Children.Length; i += 2) {
-				var keyToken = token.Children[i];
-				var valueToken = token.Children[i + 1];
+				foreach(var item in token.Children) {
 				result.Add(
-					this.Deserialize(keyToken),
-					this.Deserialize(valueToken)
+					this.Deserialize(item.Key),
+					this.Deserialize(item.Value)
 				);
 			}
 			return result;
@@ -175,16 +169,14 @@ internal class TypedObjectDeserializer : ObjectDeserializer {
 		Type keyType = targetType.GenericTypeArguments[0];
 		Type valueType = targetType.GenericTypeArguments[1];
 
-		for (int i = 0; i < token.Children.Length; i += 2) {
-			var keyToken = token.Children[i];
-			var valueToken = token.Children[i + 1];
+		foreach(var item in token.Children) {
 			result.Add(
 				keyType == typeof(object)
-					? this.Deserialize(keyToken)
-					: this.Deserialize(keyToken, keyType),
+					? this.Deserialize(item.Key)
+					: this.Deserialize(item.Key, keyType),
 				valueType == typeof(object)
-					? this.Deserialize(valueToken)
-					: this.Deserialize(valueToken, valueType)
+					? this.Deserialize(item.Value)
+					: this.Deserialize(item.Value, valueType)
 			);
 		}
 		return result;